+++ server/src/main/java/io/appulse/epmd/java/server/command/server/handler/KillRequestHandler.java
@@ -28,6 +28,11 @@ class KillRequestHandler implements RequestHandler {
 
   @Override
   public void handle (@NonNull Request request) {
+    if (!request.getContext().getServerOptions().isChecks()) {
+      request.closeConnection();
+      return;
+    }
+
     request.getContext().getNodes().clear();
     request.respondAndClose(OK);
     Runtime.getRuntime().exit(1);